Modified ScheduleCTScreen
